initial reaction ui improvements
